11/requirements.txt
openai
python-dotenv
mutagen 